Support sending flash data in chunks
The previous behavior can be emulated by setting a large chunk
size. Sending in data chunks is a way to avoid overflow on serial
transport. A recommendation for serial transport is to use a chunk
size slightly smaller than the UART RX buffer size (on the PSoC
side). Cypress' bootloader_host.exe uses 25 as chunk size - let's
default to the same value here.